Correct task that reports add-ons with no help
Change HelpReportMissing to use the same logic as PendingAddOnReleases,
which takes into account the actual status of the add-ons (as declared
in the corresponding ZapAddOn.xml file).
Extract common code, so that it can be reused by HelpReportMissing and
PendingAddOnReleases.
